openface-api/requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio[client]>=5.8.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.4.0
eventlet>=0.33.0
gevent>=22.10.0
python-engineio>=4.7.0
python-dotenv>=1.0.0
dataclasses>=0.6
pathlib2>=2.3.0
